// Frontend/src/app/fechamento/models/contigencia/agendar-solicitar-fechamento-contingencia/fechamento-dto.model.ts
export class FechamentoDto {
    diaDoMes: number;
    diaDaSemana: number;
    dataDiariaFim: string;
    dataDiariaIni: string;
    dataEspecifica: string;
    indExecucaoImediata: string;
    indSomenteDiaUtil: string;
    indUltimoDiaDoMes: string;
    numeroDeMeses: number;
    periodicidadeExecucao: number;
    codTipoFechamento: number;
    indAtivo: string;
    indAplicarHaircutProcGar: string;
    percentualHaircut: string;
    indFechamentoMensal: string;
    indGerarBaseDadosFec: string;
    valAjusteDesvioPadrao: number;
    valPercentProcOutliers: number;
    valCorteOutliers: number;
    obsValCorteOutliers: string;
    multDesvioPadrao: number;
    mesContabil: number;
    anoContabil: number;
    dataPrevia: string;
    empresa: string
    indAgendRelMovAutomatico: string;
}